Port FSTRemoteDocumentCacheTests to C++ (#4136)
