+++ Firestore/core/test/firebase/firestore/local/CMakeLists.txt
@@ -18,12 +18,17 @@ cc_test(
     #index_manager_test.cc
     #leveldb_index_manager_test.mm
     leveldb_key_test.cc
+    #leveldb_remote_document_cache_test.mm
     leveldb_util_test.cc
     local_serializer_test.cc
     #memory_index_manager_test.cc
+    #memory_remote_document_cache_test.mm
     #persistence_testing.h
     #persistence_testing.mm
+    #remote_document_cache_test.h
+    #remote_document_cache_test.mm
   DEPENDS
+    GMock::GMock
     firebase_firestore_core
     firebase_firestore_local
     firebase_firestore_local_persistence_leveldb
